Add signer
