+++ module-system/module-implementations/sov-evm/Cargo.toml
@@ -44,7 +44,7 @@ reth-primitives = { workspace = true }
 reth-rpc-types = { workspace = true }
 reth-rpc = { workspace = true }
 reth-revm = { workspace = true }
-
+secp256k1 = { workspace = true }
 
 [dev-dependencies]
 sov-evm = { path = ".", features = ["smart_contracts"] }
@@ -52,7 +52,6 @@ primitive-types = "0.12.1"
 tokio = { workspace = true }
 tempfile = { workspace = true }
 bytes = { workspace = true }
-secp256k1 = { workspace = true }
 
 
 [features]
